requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
g4f==0.2.1.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
langid==1.1.6
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-telegram-bot==20.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
